implemented generic pull. Closes #246
