+++ tests/test_23_kipoi_env.py
@@ -3,7 +3,8 @@
 import subprocess
 import os
 from kipoi.utils import read_yaml
-from kipoi.cli.env import get_env_name, export_env, list_subcomponents, merge_deps
+from kipoi.cli.env import get_env_name, export_env, merge_deps
+from kipoi.remote import list_subcomponents
 
 
 def test_env_name():
